Add feature to set different on/off humidities to Sensor
This makes it possible to set a gap between turning the ventilation on and off.
For example we start the ventilation at an humidity of 60, but we keep it running
until it drops below 50. This prevents the humidity from hanging around the turn
point, causing the ventilation to be switched on/off many times after each other.